app/pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "syla-ai-envelope-app"
version = "0.1"
maintainers = [
    { name="Osetrov Maxim", email="maxim.osetrov-2016@yandex.ru"}
]
requires-python = ">= 3.12"

# Run `pip install .`
dependencies = [
    "punq",
    "fastapi",
    "SQLAlchemy",
    "asyncpg",
    "motor",
    "uuid6",
    "asyncpg",
    "aiohttp",
    "email_validator",
    "python-multipart",
    "aiosmtplib",
    "joserfc",
    "uvicorn",
    "pydantic-settings",
    "setuptools >= 61.0"
]

# Run `pip install .[lint]`
[project.optional-dependencies]
lint = [
    "black==24.8.0",
    "mypy==1.11.1"
]

[project.urls]
page = "http://localhost:8000"

[tool.black]
line-length = 120
target-version = ['py312']
include = '\.pyi?$'
# 'extend-exclude' excludes files or directories in addition to the defaults
extend-exclude = '''
# A regex preceded with ^/ will apply only to files and directories
# in the root of the project.
(
  ^/foo.py    # exclude a file named foo.py in the root of the project
  | .*_pb2.py  # exclude autogenerated Protocol Buffer files anywhere in the project
)
'''
